+++ gpt4all-chat/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Fixed
 - Fix models.json cache location ([#3052](https://github.com/nomic-ai/gpt4all/pull/3052))
+- Fix golang syntax highlighting ([#3065](https://github.com/nomic-ai/gpt4all/pull/3065))
 
 ## [3.4.0] - 2024-10-08